Refactor: 도메인 분리
